Refactor: Move SetupKey related AM code to setupkey file
